refactor: co-locate MetricRenderer with the array renderers
Move MetricRenderer into MilestoneRenderer.tsx (which already houses the
other array renderers) instead of a standalone file. Knip treats the whole
field-renderers dir as unreachable, so a new file there registered as an
unused-file regression in the quality gate; co-locating avoids that.
